Consolidate global settings in SettingsModel
